Fixes appimage not executable error
Ran into issue where the .appimage file downloaded was not
executable by the user, so information could not be extracted.
The execute flag is now set for the user before trying to extract.

Also adds StreamHandler to main package logger, with log level of
ERROR.

In the install_appimage command, the StreamHandler is set to debug
logging.

v0.0.3